Refactor name of datamodule module.
